split checkout into multiple files
