+++ internal/job/checkout.go
@@ -8,49 +8,15 @@ import (
 	"path/filepath"
 	"runtime"
 	"slices"
-	"strings"
 	"time"
 
 	"github.com/buildkite/agent/v3/internal/experiments"
 	"github.com/buildkite/agent/v3/internal/osutil"
-	"github.com/buildkite/agent/v3/internal/self"
 	"github.com/buildkite/agent/v3/internal/shell"
 	"github.com/buildkite/agent/v3/tracetools"
 	"github.com/buildkite/roko"
 	"github.com/buildkite/shellwords"
 )
-
-// configureGitCredentialHelper sets up the agent to use a git credential helper that calls the Buildkite Agent API
-// asking for a Github App token to use when cloning. This feature is turned on serverside
-func (e *Executor) configureGitCredentialHelper(ctx context.Context) error {
-	// credential.useHttpPath is a git config setting that tells git to tell the credential helper the full URL of the repo
-	// this means that we can pass the repo being cloned up to the BK API, which can then choose (or not, if it's not permitted)
-	// to return a token for that repo.
-	//
-	// This is important for the case where a user clones multiple repos in a step - ie, if we always crammed
-	// os.Getenv("BUILDKITE_REPO") into credential helper, we'd only ever get a token for the repo that the step is running
-	// in, and not for any other repos that the step might clone.
-	err := e.shell.Command("git", "config", "--global", "credential.useHttpPath", "true").Run(ctx, shell.ShowPrompt(false))
-	if err != nil {
-		return fmt.Errorf("enabling git credential.useHttpPath: %w", err)
-	}
-
-	helper := fmt.Sprintf(`%s git-credentials-helper`, self.Path(ctx))
-	err = e.shell.Command("git", "config", "--global", "credential.helper", helper).Run(ctx, shell.ShowPrompt(false))
-	if err != nil {
-		return fmt.Errorf("configuring git credential.helper: %w", err)
-	}
-
-	return nil
-}
-
-// Disables SSH keyscan and configures git to use HTTPS instead of SSH for github.
-// We may later expand this for other SCMs.
-func (e *Executor) configureHTTPSInsteadOfSSH(ctx context.Context) error {
-	return e.shell.Command(
-		"git", "config", "--global", "url.https://github.com/.insteadOf", "git@github.com:",
-	).Run(ctx, shell.ShowPrompt(false))
-}
 
 func (e *Executor) removeCheckoutDir() error {
 	checkoutPath, _ := e.shell.Env.Get("BUILDKITE_BUILD_CHECKOUT_PATH")
@@ -394,479 +360,6 @@ func (e *Executor) checkout(ctx context.Context) error {
 	return nil
 }
 
-func hasGitSubmodules(sh *shell.Shell) bool {
-	return osutil.FileExists(filepath.Join(sh.Getwd(), ".gitmodules"))
-}
-
-func hasGitCommit(ctx context.Context, sh *shell.Shell, gitDir, commit string) bool {
-	// Resolve commit to an actual commit object
-	output, err := sh.Command("git", "--git-dir", gitDir, "rev-parse", commit+"^{commit}").RunAndCaptureStdout(ctx, shell.ShowStderr(false))
-	if err != nil {
-		return false
-	}
-
-	// Filter out commitish things like HEAD et al
-	if strings.TrimSpace(output) != commit {
-		return false
-	}
-
-	// Otherwise it's a commit in the repo
-	return true
-}
-
-// updateGitMirror clones a new git mirror (git clone --mirror ...), or updates
-// an existing git mirror to ensure relevant refs are available. It returns a
-// directory path that a checkout can use for the --reference flag. If clean
-// checkouts are enabled, dir will be a path to a snapshot of the mirror,
-// otherwise it will be the mirror.
-//
-// For efficiency reasons, updating an existing mirror is done by fetching
-// specific refspecs rather than using `git remote update` to fetch everything
-// (see https://github.com/buildkite/agent/pull/1112).
-func (e *Executor) updateGitMirror(ctx context.Context, repository string) (dir string, finalErr error) {
-	// Create a unique directory for the repository mirror
-	mirrorDir := filepath.Join(e.GitMirrorsPath, dirForRepository(repository))
-	isMainRepository := repository == e.Repository
-
-	// Create the mirrors path if it doesn't exist
-	if baseDir := filepath.Dir(mirrorDir); !osutil.FileExists(baseDir) {
-		e.shell.Commentf("Creating \"%s\"", baseDir)
-		// Actual file permissions will be reduced by umask, and won't be 0o777 unless the user has manually changed the umask to 000
-		if err := os.MkdirAll(baseDir, 0o777); err != nil {
-			return "", err
-		}
-	}
-
-	if err := e.shell.Chdir(e.GitMirrorsPath); err != nil {
-		return "", fmt.Errorf("failed to change directory to %q: %w", e.GitMirrorsPath, err)
-	}
-
-	lockTimeout := time.Second * time.Duration(e.GitMirrorsLockTimeout)
-
-	if e.Debug {
-		e.shell.Commentf("Acquiring mirror repository clone lock")
-	}
-
-	// Lock the mirror dir to prevent concurrent clones
-	cloneCtx, canc := context.WithTimeout(ctx, lockTimeout)
-	defer canc()
-	mirrorCloneLock, err := e.shell.LockFile(cloneCtx, mirrorDir+".clonelock")
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return "", ErrTimedOutAcquiringLock{Name: "clone", Err: err}
-		}
-		return "", fmt.Errorf("unable to acquire clone lock: %w", err)
-	}
-	defer mirrorCloneLock.Unlock() //nolint:errcheck // Best-effort cleanup - primary unlock checked below.
-
-	// If we don't have a mirror, we need to clone it
-	if !osutil.FileExists(mirrorDir) {
-		e.shell.Commentf("Cloning a mirror of the repository to %q", mirrorDir)
-		flags := []string{"--mirror"} // note --mirror implies --bare
-		mirrorFlags, err := shellwords.Split(e.GitCloneMirrorFlags)
-		if err != nil {
-			e.shell.Errorf("Invalid --git-clone-mirror-flags %q (%s)", e.GitCloneMirrorFlags, err)
-			return "", err
-		}
-		flags = append(flags, mirrorFlags...)
-		if err := gitClone(ctx, e.shell, flags, repository, mirrorDir); err != nil {
-			e.shell.Commentf("Removing mirror dir %q due to failed clone", mirrorDir)
-			if err := os.RemoveAll(mirrorDir); err != nil {
-				e.shell.Errorf("Failed to remove %q (%s)", mirrorDir, err)
-			}
-			return "", err
-		}
-
-		return e.snapshotMirror(ctx, repository, mirrorDir)
-	}
-
-	// If it exists, immediately release the clone lock.
-	if err := mirrorCloneLock.Unlock(); err != nil {
-		return "", fmt.Errorf("unable to release clone lock: %w", err)
-	}
-
-	if e.Debug {
-		e.shell.Commentf("Acquiring mirror repository update lock")
-	}
-
-	// Lock the mirror dir to prevent concurrent updates
-	updateCtx, canc := context.WithTimeout(ctx, lockTimeout)
-	defer canc()
-	mirrorUpdateLock, err := e.shell.LockFile(updateCtx, mirrorDir+".updatelock")
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return "", ErrTimedOutAcquiringLock{Name: "update", Err: err}
-		}
-		return "", fmt.Errorf("unable to acquire update lock: %w", err)
-	}
-	defer func() {
-		if err := mirrorUpdateLock.Unlock(); err != nil {
-			finalErr = errors.Join(finalErr, fmt.Errorf("unable to release update lock: %w", err))
-		}
-	}()
-	if isMainRepository {
-		// Check again after we get a lock, in case the other process has already updated
-		if hasGitCommit(ctx, e.shell, mirrorDir, e.Commit) {
-			e.shell.Commentf("Commit %q exists in mirror", e.Commit)
-			return e.snapshotMirror(ctx, repository, mirrorDir)
-		}
-	}
-
-	e.shell.Commentf("Updating existing repository mirror to find commit %s", e.Commit)
-
-	// Update the origin of the repository so we can gracefully handle
-	// repository renames.
-	urlChanged, err := e.updateRemoteURL(ctx, mirrorDir, repository)
-	if err != nil {
-		return "", fmt.Errorf("setting remote URL: %w", err)
-	}
-
-	if isMainRepository {
-		var refspecs []string
-		var retry bool
-
-		switch {
-		case e.RefSpec != "":
-			// If a custom refspec is provided, use it instead of the branch
-			e.shell.Commentf("Fetching and mirroring custom refspec %s", e.RefSpec)
-			refspecs = []string{e.RefSpec}
-		case e.PullRequest != "false" && strings.Contains(e.PipelineProvider, "github"):
-			e.shell.Commentf("Fetching and mirroring pull request head from GitHub. This will be retried if it fails, as the pull request head might not be available yet — GitHub creates them asynchronously")
-			var refspec string
-			if e.PullRequestUsingMergeRefspec {
-				refspec = fmt.Sprintf("refs/pull/%s/merge", e.PullRequest)
-			} else {
-				refspec = fmt.Sprintf("refs/pull/%s/head", e.PullRequest)
-			}
-			refspecs = []string{refspec}
-			retry = true
-		default:
-			// Fetch the build branch from the upstream repository into the mirror.
-			refspecs = []string{e.Branch}
-		}
-
-		// Fetch the refspecs from the upstream repository into the mirror.
-		if err := gitFetch(ctx, gitFetchArgs{
-			Shell:      e.shell,
-			GitFlags:   fmt.Sprintf("--git-dir=%s", mirrorDir),
-			Repository: "origin",
-			RefSpecs:   refspecs,
-			Retry:      retry,
-		}); err != nil {
-			return "", err
-		}
-	} else { // not the main repo.
-
-		// This is a mirror of a submodule.
-		// Update without specifying particular ref, since we don't know which
-		// ref is needed for the main build.
-		// (If it doesn't contain the needed ref, then the build would fail on
-		// a clean host or with a clean checkout.)
-		// TODO: Investigate getting the ref from the main repo and passing
-		// that in here.
-		cmd := e.shell.Command("git", "--git-dir", mirrorDir, "fetch", "origin")
-		if err := cmd.Run(ctx); err != nil {
-			return "", err
-		}
-	}
-
-	if urlChanged {
-		// Let's opportunistically fsck and gc.
-		// 1. In case of remote URL confusion (bug introduced in #1959), and
-		// 2. There's possibly some object churn when remotes are renamed.
-		if err := e.shell.Command("git", "--git-dir", mirrorDir, "fsck").Run(ctx); err != nil {
-			e.shell.Warningf("Couldn't run git fsck: %v", err)
-		}
-		if err := e.shell.Command("git", "--git-dir", mirrorDir, "gc").Run(ctx); err != nil {
-			e.shell.Warningf("Couldn't run git gc: %v", err)
-		}
-	}
-
-	return e.snapshotMirror(ctx, repository, mirrorDir)
-}
-
-// snapshotMirror creates a snapshot of the mirror. It returns the directory for
-// the rest of the checkout to use as --reference, which will be the path to a
-// snapshot, unless clean checkout is disabled, in which case it will simply
-// return mirrorDir.
-//
-// This "snapshot" is a clone of the *mirror* in a nearby directory, but on a
-// filesystem that supports hardlinks (most modern filesystems), the git objects
-// in this clone will be hardlinks into the mirror - quick to create and taking
-// up negligible extra space. Doing this ensures that any object changes in the
-// mirror (due to, say, git gc) won't corrupt the downstream reference clone
-// (the checkout). When the mirror updates, it may write new object files and
-// unlink old object files, but the snapshot continues to have access to the old
-// files via its hardlinks. (At that point, the snapshot takes up more space.)
-//
-// Like the (clean) checkout, the snapshot only needs to exist as long as the
-// current job, but we specifically remove the snapshot at the end of the job,
-// since the filesystem containing mirrors (typically) persists across jobs -
-// leaving them will let them accumulate and drift from the mirror, taking up
-// space. (In contrast, the checkout may cease to exist if the agent is
-// ephemeral, so cleaning up the checkout at the end of the job might or might
-// not be wasted effort.) Ephemeral agents can flag that they don't keep
-// their checkouts by just enabling clean checkout.
-//
-// Why no snapshots if clean checkout is disabled:
-// Disabling clean checkout is how checkouts are reused (for efficiency), so
-// the checkout can exist with its dependency on the snapshot forever, so
-// we can't delete the snapshot (otherwise we will corrupt the checkout), so
-// the mirror volume will gradually fill up with snapshots.
-// And while the mirror is updated with new objects in each job, the snapshots
-// are not updated, so the non-clean checkout will probably redundantly fetch
-// its update from the remote, so we would need extra logic to disable the
-// mirror update unless the checkout turns out to be a fresh clone.
-// It's not impossible (perhaps we need a process to age-out existing checkouts)
-// but something to think about, and when we have a good implementation enable
-// it for more cases.
-//
-// Why no snapshots if the command phase isn't included:
-// The cleanup mechanism happens when this instance of the executor tears down,
-// not when the last executor among many tears down. In a split-phase setup
-// (such as in agent-stack-k8s) where one container runs the checkout phase and
-// another runs the command phase, the snapshot would be deleted after the
-// checkout phase, which could break many git operations in the command phase.
-// Presently we have no way to pass cleanup instructions between containers,
-// which would enable this case.
-func (e *Executor) snapshotMirror(ctx context.Context, repository, mirrorDir string) (string, error) {
-	if !e.CleanCheckout || !e.includePhase("command") {
-		return mirrorDir, nil
-	}
-
-	snapshotBaseDir := filepath.Join(e.GitMirrorsPath, "snapshots")
-
-	// Create the snapshots base dir if it doesn't exist
-	if !osutil.FileExists(snapshotBaseDir) {
-		e.shell.Commentf("Creating %q", snapshotBaseDir)
-		// See comment above about umask
-		if err := os.MkdirAll(snapshotBaseDir, 0o777); err != nil {
-			return "", fmt.Errorf("creating base directory for snapshots: %w", err)
-		}
-	}
-
-	// Create a unique directory for this snapshot.
-	// MkdirTemp ensures the new dir won't collide with other agents.
-	snapshotDir, err := os.MkdirTemp(snapshotBaseDir, dirForRepository(repository))
-	if err != nil {
-		return "", fmt.Errorf("creating snapshot directory: %w", err)
-	}
-	if err := os.Chmod(snapshotDir, 0o777&^osutil.Umask); err != nil {
-		return "", fmt.Errorf("changing permissions on snapshot directory: %w", err)
-	}
-
-	// Automatically remove it during teardown
-	e.cleanupDirs = append(e.cleanupDirs, snapshotDir)
-
-	// Finally, clone the snapshot. Yes, it's a --mirror of a --mirror.
-	e.shell.Commentf("Creating mirror snapshot in %q", snapshotDir)
-	if err := gitClone(ctx, e.shell, []string{"--mirror"}, mirrorDir, snapshotDir); err != nil {
-		return "", err
-	}
-
-	return snapshotDir, nil
-}
-
-type ErrTimedOutAcquiringLock struct {
-	Name string
-	Err  error
-}
-
-func (e ErrTimedOutAcquiringLock) Error() string {
-	return fmt.Sprintf("timed out acquiring %s lock: %v", e.Name, e.Err)
-}
-
-func (e ErrTimedOutAcquiringLock) Unwrap() error { return e.Err }
-
-// updateRemoteURL updates the URL for 'origin' and reports whether the
-// URL changed from something else. If gitDir == "", it assumes the
-// local repo is in the current directory, otherwise it includes --git-dir.
-// If the remote has changed, it logs some extra information.
-func (e *Executor) updateRemoteURL(ctx context.Context, gitDir, repository string) (bool, error) {
-	// Update the origin of the repository so we can gracefully handle
-	// repository renames.
-
-	// First check what the existing remote is, for both logging and debugging
-	// purposes.
-
-	// Check if there are multiple URLs configured (e.g., via git remote set-url --add).
-	args := []string{"config", "--get-all", "remote.origin.url"}
-	if gitDir != "" {
-		args = append([]string{"--git-dir", gitDir}, args...)
-	}
-	allURLs, err := e.shell.Command("git", args...).RunAndCaptureStdout(ctx)
-	if err != nil {
-		return false, err
-	}
-
-	var gotURL string
-	urls := strings.Split(strings.TrimSpace(allURLs), "\n")
-	if len(urls) > 1 {
-		// Multiple URLs configured - fall back to git remote get-url which
-		// handles this correctly (returns primary fetch URL).
-		args = []string{"remote", "get-url", "origin"}
-		if gitDir != "" {
-			args = append([]string{"--git-dir", gitDir}, args...)
-		}
-		gotURL, err = e.shell.Command("git", args...).RunAndCaptureStdout(ctx)
-		if err != nil {
-			return false, err
-		}
-	} else {
-		// Single URL - use config output directly to avoid insteadOf transformation.
-		gotURL = urls[0]
-	}
-
-	if gotURL == repository {
-		// No need to update anything
-		return false, nil
-	}
-
-	gd := gitDir
-	if gd == "" {
-		gd = e.shell.Getwd()
-	}
-
-	e.shell.Commentf("Remote URL for git directory %s has changed (%s -> %s)!", gd, gotURL, repository)
-	e.shell.Commentf("This is usually because the repository has been renamed.")
-	e.shell.Commentf("If this is unexpected, you may see failures.")
-
-	args = []string{"remote", "set-url", "origin", repository}
-	if gitDir != "" {
-		args = append([]string{"--git-dir", gitDir}, args...)
-	}
-	return true, e.shell.Command("git", args...).Run(ctx)
-}
-
-func (e *Executor) getOrUpdateMirrorDir(ctx context.Context, repository string) (string, error) {
-	var mirrorDir string
-	// Skip updating the Git mirror before using it?
-	if e.GitMirrorsSkipUpdate {
-		mirrorDir = filepath.Join(e.GitMirrorsPath, dirForRepository(repository))
-		e.shell.Commentf("Skipping update and using existing mirror for repository %s at %s.", repository, mirrorDir)
-
-		// Check if specified mirrorDir exists, otherwise the clone will fail.
-		if !osutil.FileExists(mirrorDir) {
-			// Fall back to a clean clone, rather than failing the clone and therefore the build
-			e.shell.Commentf("No existing mirror found for repository %s at %s.", repository, mirrorDir)
-			mirrorDir = ""
-		}
-
-		// If git mirror updates are skipped, we assume there's no change
-		// to the mirror objects, so no need for snapshotting.
-		return mirrorDir, nil
-	}
-
-	return e.updateGitMirror(ctx, repository)
-}
-
-// fetchSource fetches the git source for the job. If GitSkipFetchExistingCommits is
-// enabled and the commit already exists locally, the fetch is skipped entirely.
-// When addBloblessFilter is true, --filter=blob:none is prepended to the fetch
-// flags — the caller decides based on sparse-checkout state and user-supplied
-// filters.
-func (e *Executor) fetchSource(ctx context.Context, addBloblessFilter bool) error {
-	// If configured, skip the fetch when the commit already exists locally.
-	// This is useful when a pre-populated git mirror is used with --reference,
-	// as the commit objects are already reachable and fetching is redundant.
-	if e.GitSkipFetchExistingCommits && e.Commit != "HEAD" &&
-		hasGitCommit(ctx, e.shell, ".git", e.Commit) {
-		e.shell.Commentf("Commit %q already exists locally, skipping fetch", e.Commit)
-		return nil
-	}
-
-	gitFetchFlags := e.GitFetchFlags
-	if addBloblessFilter {
-		gitFetchFlags = "--filter=blob:none " + gitFetchFlags
-	}
-
-	switch {
-	case e.RefSpec != "":
-		// If a refspec is provided then use it instead.
-		// For example, `refs/not/a/head`
-		e.shell.Commentf("Fetch and checkout custom refspec")
-		if err := gitFetch(ctx, gitFetchArgs{
-			Shell:         e.shell,
-			GitFetchFlags: gitFetchFlags,
-			Repository:    "origin",
-			RefSpecs:      []string{e.RefSpec},
-		}); err != nil {
-			return fmt.Errorf("fetching refspec %q: %w", e.RefSpec, err)
-		}
-
-	case e.PullRequest != "false" && strings.Contains(e.PipelineProvider, "github"):
-		var refspec string
-		var retry bool
-
-		if e.PullRequestUsingMergeRefspec {
-			// Merge refspecs represents a speculative merge of the PR branch against the base branch.
-			// Checking out this refspec enables testing the result of the merge before it happens.
-			// If a merge conflict exists, this refspec won't be created and the fetch will fail. In this
-			// case we want the job to fail earlier, rather than retrying the fetch (which adds ~2-3 mins job run time before failing)
-			// Note: An outer retry loop will still retry the failed checkout 3 times before failing.
-			e.shell.Commentf("Fetch and checkout pull request merge commit from GitHub")
-			retry = false
-			refspec = fmt.Sprintf("refs/pull/%s/merge", e.PullRequest)
-		} else {
-			// GitHub has a special ref which lets us fetch a pull request head, whether
-			// or not it's a current head in this repository or a fork. See:
-			// https://help.github.com/articles/checking-out-pull-requests-locally/#modifying-an-inactive-pull-request-locally
-			e.shell.Commentf("Fetch and checkout pull request head from GitHub")
-			retry = true
-			refspec = fmt.Sprintf("refs/pull/%s/head", e.PullRequest)
-		}
-		refspecs := []string{refspec}
-
-		if e.Commit == "HEAD" {
-			// If we don't know the commit, we don't want to fetch with a fallback (otherwise FETCH_HEAD
-			// will resolve during a fallback to the alphabetically earliest branch/tag - rather than the
-			// correct commit for this build)
-			if err := gitFetch(ctx, gitFetchArgs{
-				Shell:         e.shell,
-				GitFetchFlags: gitFetchFlags,
-				Repository:    "origin",
-				Retry:         retry,
-				RefSpecs:      refspecs,
-			}); err != nil {
-				return fmt.Errorf("fetching PR refspec %q: %w", refspecs, err)
-			}
-		} else {
-			// If we know the commit, also fetch it directly. The commit might not be in the history of `refspec` if there
-			// have been force pushes to the pull request, so this ensures we have it.
-			// Note: this is the typical case e.Commit != HEAD.
-			refspecs = append(refspecs, e.Commit)
-			// We aim to eliminate network round-trip as much as possible so we use a single git fetch here.
-			if err := gitFetchWithFallback(ctx, e.shell, gitFetchFlags, refspecs...); err != nil {
-				return fmt.Errorf("fetching PR refspec %q: %w", refspecs, err)
-			}
-		}
-
-		gitFetchHead, _ := e.shell.Command("git", "rev-parse", "FETCH_HEAD").RunAndCaptureStdout(ctx)
-		e.shell.Commentf("FETCH_HEAD is now `%s`", gitFetchHead)
-
-	case e.Commit == "HEAD":
-		// If the commit is "HEAD" then we can't do a commit-specific fetch and will
-		// need to fetch the remote head and checkout the fetched head explicitly.
-		e.shell.Commentf("Fetch and checkout remote branch HEAD commit")
-		if err := gitFetch(ctx, gitFetchArgs{
-			Shell:         e.shell,
-			GitFetchFlags: gitFetchFlags,
-			Repository:    "origin",
-			RefSpecs:      []string{e.Branch},
-		}); err != nil {
-			return fmt.Errorf("fetching branch %q: %w", e.Branch, err)
-		}
-
-	default:
-		// Otherwise fetch and checkout the commit directly.
-		if err := gitFetchWithFallback(ctx, e.shell, gitFetchFlags, e.Commit); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 // defaultCheckoutPhase is called by the CheckoutPhase if no global or plugin checkout
 // hook exists. It performs the default checkout on the Repository provided in the config
 func (e *Executor) defaultCheckoutPhase(ctx context.Context) (retErr error) {
@@ -1195,256 +688,5 @@ func (e *Executor) defaultCheckoutPhase(ctx context.Context) (retErr error) {
 		e.resolveCommit(ctx)
 	}
 
-	return nil
-}
-
-// prepareGitSSHKey materialises the configured GitSSHKey into a private
-// directory next to the build checkout and points GIT_SSH_COMMAND at it for
-// the duration of the checkout phase. It returns the path to the key file, a
-// cleanup function that removes the key and restores any previous
-// GIT_SSH_COMMAND, and an error. If no key is configured both the path and
-// cleanup are zero.
-//
-// Only the default checkout phase invokes this; custom checkout hooks must
-// arrange their own credentials.
-func (e *Executor) prepareGitSSHKey() (sshKeyPath string, cleanup func() error, retErr error) {
-	if e.GitSSHKey == "" {
-		return "", nil, nil
-	}
-
-	checkoutPath, exists := e.shell.Env.Get("BUILDKITE_BUILD_CHECKOUT_PATH")
-	if !exists || checkoutPath == "" {
-		return "", nil, errors.New("BUILDKITE_BUILD_CHECKOUT_PATH is not set")
-	}
-
-	parentDir := filepath.Dir(checkoutPath)
-	if err := os.MkdirAll(parentDir, 0o700); err != nil {
-		return "", nil, fmt.Errorf("creating ssh key parent directory %q: %w", parentDir, err)
-	}
-
-	pattern := ".buildkite-ssh-key-"
-	if e.PipelineSlug != "" {
-		pattern += badCharsRE.ReplaceAllString(e.PipelineSlug, "-") + "-"
-	}
-
-	// os.MkdirTemp creates the directory with mode 0o700 on Unix, keeping the
-	// key file in its own private directory instead of beside the checkout.
-	sshKeyDir, err := os.MkdirTemp(parentDir, pattern)
-	if err != nil {
-		return "", nil, fmt.Errorf("creating ssh key directory: %w", err)
-	}
-	defer func() {
-		if retErr != nil {
-			_ = os.RemoveAll(sshKeyDir)
-		}
-	}()
-
-	sshKeyPath = filepath.Join(sshKeyDir, "id")
-	// Most SSH key parsers require a trailing newline; tolerate either form
-	// of input and always write a single one.
-	keyBytes := []byte(strings.TrimRight(e.GitSSHKey, "\n") + "\n")
-	if err := os.WriteFile(sshKeyPath, keyBytes, 0o600); err != nil {
-		return "", nil, fmt.Errorf("writing ssh key file: %w", err)
-	}
-
-	previousGitSSHCommand, hadPreviousGitSSHCommand := e.shell.Env.Get("GIT_SSH_COMMAND")
-	e.shell.Env.Set("GIT_SSH_COMMAND", gitSSHCommandForKeyFile(sshKeyPath, previousGitSSHCommand))
-
-	cleanup = func() error {
-		if hadPreviousGitSSHCommand {
-			e.shell.Env.Set("GIT_SSH_COMMAND", previousGitSSHCommand)
-		} else {
-			e.shell.Env.Remove("GIT_SSH_COMMAND")
-		}
-		if err := os.RemoveAll(sshKeyDir); err != nil && !os.IsNotExist(err) {
-			return err
-		}
-		return nil
-	}
-
-	return sshKeyPath, cleanup, nil
-}
-
-func gitSSHCommandForKeyFile(path, previous string) string {
-	keyOptions := fmt.Sprintf("-i %s -o IdentitiesOnly=yes", shellwords.Quote(path))
-	if previous == "" {
-		return "ssh " + keyOptions
-	}
-	return strings.TrimSpace(previous) + " " + keyOptions
-}
-
-// gitFetchWithFallback run git fetch for refspecs, when it fails on recoverable reason, it will retry fetching
-// all heads and refs.
-func gitFetchWithFallback(ctx context.Context, shell *shell.Shell, gitFetchFlags string, refspecs ...string) error {
-	if len(refspecs) == 0 {
-		return fmt.Errorf("no refspecs provided for git fetch")
-	}
-
-	// Try to fetch all refspecs in a single call first
-	err := gitFetch(ctx, gitFetchArgs{
-		Shell:         shell,
-		GitFetchFlags: gitFetchFlags,
-		Repository:    "origin",
-		RefSpecs:      refspecs,
-	})
-	if err == nil {
-		return nil // all refspecs worked in single fetch
-	}
-
-	if gerr := new(gitError); errors.As(err, &gerr) {
-		switch gerr.Type {
-		case gitErrorFetchBadReference:
-			// refspecs might contains short SHA
-			break
-		default:
-			// bail due to repository corruption or other non recoverable issue
-			return fmt.Errorf("fetching refspecs %v: %w", refspecs, err)
-		}
-	}
-
-	// The refspecs might be something that's not possible to fetch directly
-	// (e.g. short commit hashes), so we fall back to fetching all heads and tags,
-	// hoping that the refspecs are included.
-	shell.Commentf("Some refspec fetches failed, trying to fetch all heads and tags")
-	// By default `git fetch origin` will only fetch tags which are
-	// reachable from a fetches branch. git 1.9.0+ changed `--tags` to
-	// fetch all tags in addition to the default refspec, but pre 1.9.0 it
-	// excludes the default refspec.
-	gitFetchRefspec, err := shell.Command("git", "config", "remote.origin.fetch").RunAndCaptureStdout(ctx)
-	if err != nil {
-		return fmt.Errorf("getting remote.origin.fetch: %w", err)
-	}
-
-	if err := gitFetch(ctx, gitFetchArgs{
-		Shell:         shell,
-		GitFetchFlags: gitFetchFlags,
-		Repository:    "origin",
-		Retry:         true,
-		RefSpecs:      []string{gitFetchRefspec, "+refs/tags/*:refs/tags/*"},
-	}); err != nil {
-		return fmt.Errorf("fetching refspecs %v: %w", refspecs, err)
-	}
-
-	return nil
-}
-
-const CommitMetadataKey = "buildkite:git:commit"
-
-// sendCommitToBuildkite sends commit information (commit, author, subject, body) to Buildkite, as the BK backend doesn't
-// have access to user's VCSes. To do this, we set a special meta-data key in the build, but only if it isn't already present
-// Functionally, this means that the first job in a build (usually a pipeline upload or similar) will push the commit info
-// to buildkite, which uses this info to display commit info in the UI eg in the title for the build
-// note that we bail early if the key already exists, as we don't want to overwrite it
-func (e *Executor) sendCommitToBuildkite(ctx context.Context) error {
-	if e.SkipCheckout {
-		return nil
-	}
-
-	e.shell.Commentf("Checking to see if git commit information needs to be sent to Buildkite...")
-
-	commitResolved, _ := e.shell.Env.Get("BUILDKITE_COMMIT_RESOLVED")
-	if commitResolved == "true" {
-		// we can skip the metadata shenanigans here and push straight through
-		e.shell.Commentf("BUILDKITE_COMMIT is already resolved and meta-data populated, skipping")
-		return nil
-	}
-
-	cmd := e.shell.Command(self.Path(ctx), "meta-data", "exists", CommitMetadataKey)
-	if err := cmd.Run(ctx); err == nil {
-		// Command exited 0, ie the key exists, so we don't need to send it again
-		e.shell.Commentf("Git commit information has already been sent to Buildkite")
-		return nil
-	}
-
-	e.shell.Commentf("Sending Git commit information back to Buildkite")
-	// Format:
-	//
-	// commit 0123456789abcdef0123456789abcdef01234567
-	// abbrev-commit 0123456789
-	// Author: John Citizen <john@example.com>
-	//
-	//    Subject of the commit message
-	//
-	//    Body of the commit message, which
-	//    may span multiple lines.
-	gitArgs := []string{
-		"--no-pager",
-		"log",
-		"-1",
-		e.Commit,
-		"-s", // --no-patch was introduced in v1.8.4 in 2013, but e.g. CentOS 7 isn't there yet
-		"--no-color",
-		"--format=commit %H%nabbrev-commit %h%nAuthor: %an <%ae>%n%n%w(0,4,4)%B",
-	}
-	out, err := e.shell.Command("git", gitArgs...).RunAndCaptureStdout(ctx)
-	if err != nil {
-		return fmt.Errorf("getting git commit information: %w", err)
-	}
-
-	stdin := strings.NewReader(out)
-	cmd = e.shell.CloneWithStdin(stdin).Command(self.Path(ctx), "meta-data", "set", CommitMetadataKey)
-	if err := cmd.Run(ctx); err != nil {
-		return fmt.Errorf("sending git commit information to Buildkite: %w", err)
-	}
-
-	return nil
-}
-
-func (e *Executor) resolveCommit(ctx context.Context) {
-	commitRef, _ := e.shell.Env.Get("BUILDKITE_COMMIT")
-	if commitRef == "" {
-		e.shell.Warningf("BUILDKITE_COMMIT was empty")
-		return
-	}
-	cmdOut, err := e.shell.Command("git", "rev-parse", commitRef).RunAndCaptureStdout(ctx)
-	if err != nil {
-		e.shell.Warningf("Error running git rev-parse %q: %v", commitRef, err)
-		return
-	}
-	trimmedCmdOut := strings.TrimSpace(string(cmdOut))
-	if trimmedCmdOut != commitRef {
-		e.shell.Commentf("Updating BUILDKITE_COMMIT from %q to %q", commitRef, trimmedCmdOut)
-		e.shell.Env.Set("BUILDKITE_COMMIT", trimmedCmdOut)
-	}
-}
-
-// This is the same thing that git does at the end of clone when it is
-// passed --dissociate:
-// https://github.com/git/git/blob/6e8d538aab8fe4dd07ba9fb87b5c7edcfa5706ad/builtin/clone.c#L843-L859
-// This kind of git surgery is acceptable, because it's how one would dissociate
-// a reference clone prior to Git 2.3.
-func (e *Executor) dissociateIfNeeded(ctx context.Context, gitDir string) error {
-	alternates := filepath.Join(gitDir, "objects", "info", "alternates")
-	if !osutil.FileExists(alternates) {
-		return nil
-	}
-	e.shell.Commentf("Dissociating existing reference clone because git mirror checkout mode is %q", e.GitMirrorCheckoutMode)
-	if err := gitRepack(ctx, e.shell, "-a", "-d"); err != nil {
-		return fmt.Errorf("cleaning up reference clone: %w", err)
-	}
-	if err := os.Remove(alternates); err != nil {
-		return fmt.Errorf("removing alternates file: %w", err)
-	}
-	return nil
-}
-
-// reassociateIfNeeded writes a new alternates file into gitDir/objects/info,
-// referring to mirrorDir/objects. This allows the repo in gitDir to reuse
-// objects from mirrorDir, but at the risk of those objects becoming unavailable
-// later on.
-func (e *Executor) reassociateIfNeeded(ctx context.Context, gitDir, mirrorDir string) error {
-	alternates := filepath.Join(gitDir, "objects", "info", "alternates")
-	if osutil.FileExists(alternates) {
-		return nil
-	}
-	e.shell.Commentf("Re-associating existing clone because git mirror checkout mode is %q", e.GitMirrorCheckoutMode)
-	objects := filepath.Join(mirrorDir, "objects")
-	if !osutil.FileExists(objects) {
-		return fmt.Errorf("objects directory missing from mirror directory %s", mirrorDir)
-	}
-	objects += "\n"
-	if err := os.WriteFile(alternates, []byte(objects), 0o644); err != nil {
-		return fmt.Errorf("writing alternates file: %w", err)
-	}
 	return nil
 }
